feat: rename delete and leave modal folder
