+++ apps/web/src/routes/my-kiwis/-modals/index.ts
@@ -1,4 +1,4 @@
 export { default as DetailKiwiModal } from "./detail-kiwi";
 export { default as CreateKiwiModal } from "./create-kiwi";
-export { default as DeleteKiwiModal } from "./delete-modal";
-export { default as LeaveKiwiModal } from "./leave-modal";
+export { default as DeleteKiwiModal } from "./delete-kiwi";
+export { default as LeaveKiwiModal } from "./leave-kiwi";
